+++ src/main/java/org/octri/fhir_sandbox/service/SandboxService.java
@@ -1,12 +1,10 @@
 package org.octri.fhir_sandbox.service;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.CodeType;
 import org.hl7.fhir.r4.model.IntegerType;
 import org.hl7.fhir.r4.model.Parameters;
@@ -14,6 +12,7 @@ import org.hl7.fhir.r4.model.StringType;
 import org.octri.authentication.server.security.entity.User;
 import org.octri.fhir_sandbox.config.FhirServerProperties;
 import org.octri.fhir_sandbox.domain.Sandbox;
+import org.octri.fhir_sandbox.domain.SandboxStatus;
 import org.octri.fhir_sandbox.domain.SmartClient;
 import org.octri.fhir_sandbox.repository.SandboxRepository;
 import org.octri.fhir_sandbox.repository.SmartClientRepository;
@@ -22,7 +21,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.rest.server.exceptions.BaseServerResponseException;
 
 /**
  * Business logic for working with {@link Sandbox} entities.
@@ -56,6 +54,18 @@ public class SandboxService {
 	 */
 	public Iterable<Sandbox> findAll() {
 		return repository.findAll();
+	}
+
+	/**
+	 * Finds sandboxes with the specified status (for all owners)
+	 * 
+	 * TODO: Method security to ensure only admins can view all sandboxes.
+	 * 
+	 * @param status
+	 * @return
+	 */
+	public List<Sandbox> findByStatus(SandboxStatus status) {
+		return repository.findByStatus(status);
 	}
 
 	/**
@@ -113,13 +123,8 @@ public class SandboxService {
 	 */
 	public Sandbox createSandbox(Sandbox sandbox) {
 		createPartitionForSandbox(sandbox);
+		sandbox.setStatus(SandboxStatus.INITIALIZING);
 		Sandbox savedSandbox = repository.save(sandbox);
-		// TODO: relocate call to loadSampleData so that it can be optional
-		try {
-			loadSampleData(savedSandbox);
-		} catch (IOException e) {
-			log.error("Problem encountered reading sample data resources", e);
-		}
 		return savedSandbox;
 	}
 
@@ -147,27 +152,28 @@ public class SandboxService {
 	}
 
 	/**
-	 * Loads sample FHIR resources then posts them to the FHIR server
+	 * Asynchronously performs setup tasks for sandboxes
+	 * 
+	 * First sets the status to INITIALIZING, then decides whether the load
+	 * sample FHIR resources before setting the status to READY.
+	 * 
+	 * If loadSampleData throws an error, the status is instead updated to ERROR
 	 * 
 	 * @param sandbox
+	 * @param importSampleData
 	 */
 	@Async
-	public void loadSampleData(Sandbox sandbox) throws IOException {
-		var sampleData = sampleDataService.getAllSampleBundles();
-		var fhirClient = fhirContext.newRestfulGenericClient(getSandboxFhirUrl(sandbox));
-		for (var bundle : sampleData) {
-			try {
-				Bundle resp = fhirClient
-						.transaction()
-						.withBundle(bundle)
-						.execute();
-				// TODO: check outcome and handle failures
-			} catch (BaseServerResponseException e) {
-				log.error("Error response to transaction with sandbox server " + getSandboxFhirUrl(sandbox), e);
-			} catch (Error e) {
-				log.error("Error encountered during transaction with sandbox server " + getSandboxFhirUrl(sandbox), e);
+	public void initializeSandbox(Sandbox sandbox, Boolean importSampleData) {
+		try {
+			if (importSampleData) {
+				sampleDataService.loadSampleData(getSandboxFhirUrl(sandbox));
 			}
+			sandbox.setStatus(SandboxStatus.READY);
+		} catch (Exception e) {
+			log.error("Error encountered loading sample resources", e);
+			sandbox.setStatus(SandboxStatus.ERROR);
 		}
+		save(sandbox);
 	}
 
 	/**